gui_finish_main_part
Code divided  into 3 .py GUI. Added new function-get_coordinates.
