+++ mapka_Maciej.html
@@ -1,0 +1,93 @@
+<!DOCTYPE html>
+<html>
+<head>
+    
+    <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
+    
+        <script>
+            L_NO_TOUCH = false;
+            L_DISABLE_3D = false;
+        </script>
+    
+    <style>html, body {width: 100%;height: 100%;margin: 0;padding: 0;}</style>
+    <style>#map {position:absolute;top:0;bottom:0;right:0;left:0;}</style>
+    <script src="https://cdn.jsdelivr.net/npm/leaflet@1.9.3/dist/leaflet.js"></script>
+    <script src="https://code.jquery.com/jquery-3.7.1.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/Leaflet.awesome-markers/2.0.2/leaflet.awesome-markers.js"></script>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/leaflet@1.9.3/dist/leaflet.css"/>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css"/>
+    <link rel="stylesheet" href="https://netdna.bootstrapcdn.com/bootstrap/3.0.0/css/bootstrap.min.css"/>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@fortawesome/fontawesome-free@6.2.0/css/all.min.css"/>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/Leaflet.awesome-markers/2.0.2/leaflet.awesome-markers.css"/>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/python-visualization/folium/folium/templates/leaflet.awesome.rotate.min.css"/>
+    
+            <meta name="viewport" content="width=device-width,
+                initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
+            <style>
+                #map_24557d2f5b7247549ee342d172b93d3d {
+                    position: relative;
+                    width: 100.0%;
+                    height: 100.0%;
+                    left: 0.0%;
+                    top: 0.0%;
+                }
+                .leaflet-container { font-size: 1rem; }
+            </style>
+        
+</head>
+<body>
+    
+    
+            <div class="folium-map" id="map_24557d2f5b7247549ee342d172b93d3d" ></div>
+        
+</body>
+<script>
+    
+    
+            var map_24557d2f5b7247549ee342d172b93d3d = L.map(
+                "map_24557d2f5b7247549ee342d172b93d3d",
+                {
+                    center: [53.32847, 22.00774],
+                    crs: L.CRS.EPSG3857,
+                    zoom: 14,
+                    zoomControl: true,
+                    preferCanvas: false,
+                }
+            );
+
+            
+
+        
+    
+            var tile_layer_f16d3a2f323fe97f1bbc9d9171477daf = L.tileLayer(
+                "https://tile.openstreetmap.org/{z}/{x}/{y}.png",
+                {"attribution": "\u0026copy; \u003ca href=\"https://www.openstreetmap.org/copyright\"\u003eOpenStreetMap\u003c/a\u003e contributors", "detectRetina": false, "maxNativeZoom": 19, "maxZoom": 19, "minZoom": 0, "noWrap": false, "opacity": 1, "subdomains": "abc", "tms": false}
+            );
+        
+    
+            tile_layer_f16d3a2f323fe97f1bbc9d9171477daf.addTo(map_24557d2f5b7247549ee342d172b93d3d);
+        
+    
+            var marker_cfbdb87713414a7c01515c185900cc1f = L.marker(
+                [53.32847, 22.00774],
+                {}
+            ).addTo(map_24557d2f5b7247549ee342d172b93d3d);
+        
+    
+        var popup_f805ac572ec5b6a170cf2534e258231e = L.popup({"maxWidth": "100%"});
+
+        
+            
+                var html_ad5c91a23ce5ccdcf3dec09070cad0a1 = $(`<div id="html_ad5c91a23ce5ccdcf3dec09070cad0a1" style="width: 100.0%; height: 100.0%;">Pacjent: Maciej Malinowski PESEL: 00231211251</div>`)[0];
+                popup_f805ac572ec5b6a170cf2534e258231e.setContent(html_ad5c91a23ce5ccdcf3dec09070cad0a1);
+            
+        
+
+        marker_cfbdb87713414a7c01515c185900cc1f.bindPopup(popup_f805ac572ec5b6a170cf2534e258231e)
+        ;
+
+        
+    
+</script>
+</html>
